backend/requirement.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
ffmpeg-python==0.2.0
future==1.0.0
kombu==5.6.1
numpy==2.3.5
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
pydub==0.25.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
redis==7.1.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.2.14
webrtcvad==2.0.10
